Move route/util to util
Which contains global utilites shared between auth/, database/, ...
