+++ backend/app.py
@@ -5,7 +5,7 @@ from typing import Any, Mapping
 from flask import Flask
 
 from database.util import connect_database_for_app
-from route.util import fetch_page
+from util import fetch_page
 
 from auth.route import auth_bp
